add context method to return components from view
this simplifies the API and reduces the amount of EnTT that the user
interacts with directly. before the context view returned an EnTT view
that was then iterated. now, we return the structured bindings of the
view directly in the Context class and the only EnTT-specific object
that is returned is the EnTT entity id (uint32).