EnTT setup (#3)
* including EnTT library

* WIP: messing around with design of the library API

* reworking logger to be more clear

FUTURE TODO: learn how to properly hide spdlog from the user so that
just xlamb API is exposed (aka goal is for xlamb's dependencies to be
PRIVATE)

* turning off stderr logging during development

* switching to lower-case naming style for now

* returning to capitalized class naming

* changing logger to accept vcariadic args to permit fmtlib style

* basic entity controls

create, manipulate components, destroy

* planning API

* entity name helper function

* adding helper functions that return mutliple entities/components

allows for multiple component retrieval and for the use of entt views

* additional sandbox testing

* more sandbox testing

* renaming architect to simulator

* commenting our native script planning

* basic single-strain sim w/ vax works

NEXT STEPS:
- define what can be moved into native scripts
- define the API that we want to expose to users/scripts

* deleting commented-out code

* adding entity lookup by tag name

* reorganizing main.cpp

* simulator now stores setup, sim, and report as std::functions

user can define these functions and bind them to the simulator

* moving model-specific code to model.hpp

this makes main.cpp very small and simple to pre-generate

* reorganizing model code

* consistent typename variables

* removing unneeded forward declaration

* planning out wrapper class for entt views

* removing view.hpp

* add context method to return components from view

this simplifies the API and reduces the amount of EnTT that the user
interacts with directly. before the context view returned an EnTT view
that was then iterated. now, we return the structured bindings of the
view directly in the Context class and the only EnTT-specific object
that is returned is the EnTT entity id (uint32).

* minor edits

removing commented code

* reorganizing model code for clearer API planning

* basic model.hpp comments

* removing entt from linked libraries in example model

* adding basic flu model used in PR description

* adding entt license

im not sure if this is necessary since the license is copied by CMake
into the build, but id rather be safe